requirements.txt
fastapi[standard]
uvicorn
passlib[bcrypt]
python-jose
sqlalchemy
alembic
psycopg2-binary
argon2-cffi
joblib
sentence_transformers
numpy
keras
tensorflow
tf-keras